sqlite3: fix CVE-2025-7709
CVE-2025-7709:
SQLite is vulnerable to memory corruption due to an integer overflow in the FTS5 extension that
can occur when the size of an array of tombstone pointers is calculated, and that size is then
truncated in order to fit into a 32-bit integer. A remote attacker could exploit this vulnerability
in order to cause a heap buffer overflow which could result in a denial-of-service condition, or
leverage the memory corruption in order to achieve higher impacts.

References:
[https://nvd.nist.gov/vuln/detail/CVE-2025-7709]
[https://github.com/google/security-research/security/advisories/GHSA-v2c8-vqqp-hv3g]